+++ src/main/java/com/drakmyth/minecraft/manufactory/init/ModTileEntityTypes.java
@@ -6,9 +6,9 @@
 package com.drakmyth.minecraft.manufactory.init;
 
 import com.drakmyth.minecraft.manufactory.Reference;
-import com.drakmyth.minecraft.manufactory.blocks.entities.BallMillTileEntity;
-import com.drakmyth.minecraft.manufactory.blocks.entities.GrinderTileEntity;
-import com.drakmyth.minecraft.manufactory.blocks.entities.LatexCollectorTileEntity;
+import com.drakmyth.minecraft.manufactory.blocks.entities.BallMillBlockEntity;
+import com.drakmyth.minecraft.manufactory.blocks.entities.GrinderBlockEntity;
+import com.drakmyth.minecraft.manufactory.blocks.entities.LatexCollectorBlockEntity;
 
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraftforge.fmllegacy.RegistryObject;
@@ -18,7 +18,7 @@ import net.minecraftforge.registries.ForgeRegistries;
 public final class ModTileEntityTypes {
     public static final DeferredRegister<BlockEntityType<?>> TILE_ENTITY_TYPES = DeferredRegister.create(ForgeRegistries.BLOCK_ENTITIES, Reference.MOD_ID);
 
-    public static final RegistryObject<BlockEntityType<GrinderTileEntity>> GRINDER = TILE_ENTITY_TYPES.register("grinder", () -> BlockEntityType.Builder.of(GrinderTileEntity::new, ModBlocks.GRINDER.get()).build(null));
-    public static final RegistryObject<BlockEntityType<BallMillTileEntity>> BALL_MILL = TILE_ENTITY_TYPES.register("ball_mill", () -> BlockEntityType.Builder.of(BallMillTileEntity::new, ModBlocks.BALL_MILL.get()).build(null));
-    public static final RegistryObject<BlockEntityType<LatexCollectorTileEntity>> LATEX_COLLECTOR = TILE_ENTITY_TYPES.register("latex_collector", () -> BlockEntityType.Builder.of(LatexCollectorTileEntity::new, ModBlocks.LATEX_COLLECTOR.get()).build(null));
+    public static final RegistryObject<BlockEntityType<GrinderBlockEntity>> GRINDER = TILE_ENTITY_TYPES.register("grinder", () -> BlockEntityType.Builder.of(GrinderBlockEntity::new, ModBlocks.GRINDER.get()).build(null));
+    public static final RegistryObject<BlockEntityType<BallMillBlockEntity>> BALL_MILL = TILE_ENTITY_TYPES.register("ball_mill", () -> BlockEntityType.Builder.of(BallMillBlockEntity::new, ModBlocks.BALL_MILL.get()).build(null));
+    public static final RegistryObject<BlockEntityType<LatexCollectorBlockEntity>> LATEX_COLLECTOR = TILE_ENTITY_TYPES.register("latex_collector", () -> BlockEntityType.Builder.of(LatexCollectorBlockEntity::new, ModBlocks.LATEX_COLLECTOR.get()).build(null));
 }
