+++ src/main/java/com/drakmyth/minecraft/manufactory/menus/BallMillUpgradeMenu.java
@@ -6,7 +6,7 @@
 package com.drakmyth.minecraft.manufactory.menus;
 
 import com.drakmyth.minecraft.manufactory.LogMarkers;
-import com.drakmyth.minecraft.manufactory.blocks.entities.BallMillTileEntity;
+import com.drakmyth.minecraft.manufactory.blocks.entities.BallMillBlockEntity;
 import com.drakmyth.minecraft.manufactory.init.ModBlocks;
 import com.drakmyth.minecraft.manufactory.init.ModContainerTypes;
 import com.drakmyth.minecraft.manufactory.items.upgrades.IMillingBallUpgrade;
@@ -34,7 +34,7 @@ public class BallMillUpgradeMenu extends AbstractContainerMenu {
     private static final Logger LOGGER = LogManager.getLogger();
     public final ItemStackHandler upgradeInventory;
     private final ContainerLevelAccess posCallable;
-    private final BallMillTileEntity tileEntity;
+    private final BallMillBlockEntity tileEntity;
 
     public BallMillUpgradeMenu(int windowId, Inventory playerInventory, FriendlyByteBuf data) {
         this(windowId, new InvWrapper(playerInventory), playerInventory.player, data.readBlockPos());
@@ -45,7 +45,7 @@ public class BallMillUpgradeMenu extends AbstractContainerMenu {
         LOGGER.debug(LogMarkers.CONTAINER, "Initializing BallMillUpgradeContainer...");
         Level world = player.getCommandSenderWorld();
         posCallable = ContainerLevelAccess.create(world, pos);
-        tileEntity = (BallMillTileEntity)world.getBlockEntity(pos);
+        tileEntity = (BallMillBlockEntity)world.getBlockEntity(pos);
         upgradeInventory = tileEntity.getUpgradeInventory();
 
         // Ball Mill Slots
